LightVisualiser : Move from GafferSceneUI to IECoreGLPreview
