+++ python/GafferSceneUI/__init__.py
@@ -36,6 +36,7 @@
 ##########################################################################
 
 __import__( "GafferImageUI" )
+__import__( "GafferScene" )
 
 from _GafferSceneUI import *
